refactor korosensei mod
